feat(api): wire Demo 2 CQRS pipeline, middleware, DI, HTTP clients, and SSE

+++ Tripcare360.Application/Dtos/Flight/FlightStatusResponse.cs
@@ -5,7 +5,7 @@ public record FlightStatusResponse(
     DateTime DepartureDate,
     string AirlineName,
     bool IsDelayed,
-    int ActualDelayHours,
+    double ActualDelayHours,
     bool IsBaggageDelayed,
-    int BaggageDelayHours
+    double BaggageDelayHours
 );
